Correct the stated amount of Base64's data inflation. Used to say 4x; correct value is +33%.
I think back when I was doing this QR code stuff, I had binary data that should have been able to fit in the QR code, but when I tried encoding to Base64, it was too big. And, without thinking too much, I did that rough calculation on the Base64 expansion, and just accepted the 4x inflation without thinking too much about it. Then discovered this Base45 exploit, and started working on that.